Replace `force_reenumeration` with `force_reset`
This removes the `force_reenumeration` method and replaces it with an
implementation of `usb_device::bus::UsbBus::force_reset`. I'm not at all
sure if this is the right thing to do, but here's my reasoning why I
think it might be:

- Both methods seem to have the same purpose.
- I can't see a reason _not_ to implement `force_reset`
  I might be missing some vital context, of course, but at least I can
  say that this method seems to work for me (USB device is recognized by
  the host PC right after flashing the program; it isn't if I don't call
  the method).
- I can't find a way to actually call `force_reenumeration`.
  It's a method on `stm32_usbd::bus::UsbBus`, but the constructor of
  that type only provides a `UsbBusAllocator`. I couldn't figure out how
  to get a `UsbBus` instance to call that method on.

My best guess is that something changed in `usb-device`, `stm32-usbd`
was updated, and that update caused a situation that no longer makes
sense. If that's true, then the fix in this commit is probably correct.

If this is wrong and I'm missing something (which is totally possible!),
I'd appreciate some guidance on how to call `force_reenumeration`.