Updated wakeup logic to check for spurious wakeups
Currently, at least with the 32L0358DISCOVERY Devkit, starting the
stack with the USB port unplugged, or removing the USB port results in
the USB stack being left in state UsbDeviceState::Default. The
expected behavior is that the USB stack stays in the suspended state.

This causes the resume() functions to be called in the HAL, which
means it is not possible to leave the USB hardware in the low power
state.

According to ST document RM0267, section 32.5.5, it is possible to get
spurious wakeup events on the USB bus, even if the bus is still
suspended. Table 159 describes the proper response:

|(RXDP,RXDM)| Wakeup Event        | Required Action         |
|-----------|---------------------|-------------------------|
|  (0,0)    | Root Reset          | None                    |
|  (0,1)    | None (noise on bus) | Go Back in Suspend Mode |
|  (1,0)    | Root Resume         | None                    |
|  (1,1)    | N/A (Noise on bus)  | Go Back in Suspend Mode |

In summary, the status of these bits needs to be checked to make sure
that the source of the wakeup is not spurious noise, and allow the
system to go back to sleep if so.